% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perceptron.R
\name{plot_perceptron}
\alias{plot_perceptron}
\title{jk}
\usage{
plot_perceptron(ans, xlo, xhi, ylo, yhi, repper = 100)
}
\arguments{
\item{ans}{jk}

\item{xlo}{jk}

\item{xhi}{jk}

\item{ylo}{jk}

\item{yhi}{jk}

\item{repper}{jk}
}
\description{
jk
}
\details{
jk
}
\examples{
\dontrun{
p <- list(c(1, 1), c(1, 2), c(2, -1), c(2, 0), c(-1, 2), c(-2, 1), c(-1, -1), c(-2 ,-2))
t <- list(c(0, 0), c(0, 0), c(0, 1), c(0, 1), c(1, 0), c(1, 0), c(1, 1), c(1, 1))
verbose <- TRUE
W_0 <- matrix(c(1, 0, 0, 1), ncol = 2)
b_0 <- c(1, 1)
ans <- perceptron(p, t, verbose, W_0, b_0)

plot_perceptron(ans, -5, 5, -5, 5)}
}
